VideoCore/Shader: Move DebugData to a separate file
